+++ changelog.md
@@ -1,2 +1,12 @@
+# 1.1.0
+- <cg>Added</c> settings to:
+    - Control the <cy>percentage precision</c>. (How many decimal places to show)
+    - Control the <cy>label opacity</c>.
+    - Control the <cy>label size</c>.
+    - Control the <cy>label font</c>.
+    - Control the <cy>label color</c>.
+- <cy>Changed</c> the mod icon.
+- <cg>Added</c> a button in the pause menu that opens the mod's <cy>settings</c>.
+
 # 1.0.0
 - Initial release.